requirements.txt
Django==4.2.5
gunicorn==21.2.0
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.0
mysqlclient==2.2.0
pymysql==1.1.0